Add mixer: Audio mixing works using the Web Audio API.
Use a MediaStreamAudioSourceNode and a MediaElementAudioSourceNode
for the parent track audio and the new input stream source. Merge
them by routing through a merger node, that is further connected
to the default audio context destination (say, your speakers) and
a MediaStreamAudioDestinationNode. Use the stream from this
destination node to feed the mixed audio blob to a HTML audio
element to play.

Signed-off-by: Samikshan Bairagya <samikshan@gmail.com>